+++ tools/buildall-rasp.sh
@@ -9,18 +9,33 @@ fi
 echo installing development tools...
 ${SUDO} apt-get -y install tzdata
 ${SUDO} apt-get install -y gcc g++ python3 python3-dev python3-pip flex bison \
-libtool shtool automake autoconf autotools-dev make dbus libdbus-1-3 \
+libtool shtool automake autoconf autotools-dev make dbus dbus-bin libdbus-1-3 \
 libdbus-1-dev libjsoncpp-dev libkrb5-3 libkrb5-dev liblz4-1 \
 liblz4-dev openssl libssl1.1 libssl-dev libcppunit-1.14-0 libcppunit-dev \
-libfuse3-3 libfuse3-dev \
-bash net-tools procps swig default-jdk-headless cmake libcommons-cli-java ccache
+libfuse3-3 libfuse3-dev bash net-tools procps swig default-jdk-headless cmake \
+libcommons-cli-java ccache curl fuse3
 
-${SUDO} apt-get -y install sip-tools || apt-get -y install sip-dev python3-sip python3-sip-dev || true
-${SUDO} apt-get -y install libjsoncpp1 || apt-get -y install libjsoncpp25 || apt-get -y install libjsoncpp24
+echo install nodejs 18
+if ! curl -o- https://raw.githubusercontent.com/nvm-sh/nvm/v0.39.3/install.sh | bash; then exit 1; fi
+if [[ -s $HOME/.nvm/nvm.sh ]]; then source $HOME/.nvm/nvm.sh;nvm install 18;fi
+
+${SUDO} apt-get -y install sip-tools || ${SUDO} apt-get -y install sip-dev python3-sip python3-sip-dev || true
+
+for i in 1 4 5 6; do
+if ${SUDO} apt-get -y install libjsoncpp$i; then 
+    break;
+fi
+done
+if ! dpkg -l libjsoncpp$i > /dev/null; then
+    echo Error unable to install libjsoncpp
+    exit 1
+fi
+
 ${SUDO} apt-get -y install git devscripts debhelper expect screen vim
 ${SUDO} apt-get -y install python3-wheel python3-numpy || pip3 install wheel numpy
-#npm -g install assert browserify buffer exports long lz4 process put safe-buffer \
-# stream xxhash xxhashjs minify webpack webpack-cli vm events crypto-browserify \
-# stream-browserify
+
+npm -g install assert browserify buffer exports long lz4 process put safe-buffer \
+ stream xxhash xxhashjs minify webpack webpack-cli vm events crypto-browserify \
+ stream-browserify
 
 bash ./makerpcf-rasp.sh
